+++ src/com/nixinova/player/Controller.java
@@ -230,16 +230,16 @@ public class Controller {
 		// Above the ground if the block one texel beneath the player's feet is air
 		TxCoord curTx = this.pos.toTx();
 		BlockCoord blockOneTxDown = Coord3.fromTx(curTx.x, curTx.y - 1, curTx.z).toBlock();
+		boolean belowTxIsNotVoid = this.game.world.isWithinWorld(blockOneTxDown);
 		boolean belowTxIsAir = this.game.world.isAir(blockOneTxDown);
-		return belowTxIsAir;
+		return belowTxIsNotVoid && belowTxIsAir;
 	}
 
 	private boolean belowGround() {
 		// Below the ground if the block one texel above the player's feet is air
 		TxCoord footTx = this.pos.toTx();
 		BlockCoord blockOneTxUp = Coord3.fromTx(footTx.x, footTx.y + 1, footTx.z).toBlock();
-		boolean inVoid = footTx.y <= 0;
 		boolean aboveTxIsAir = this.game.world.isAir(blockOneTxUp);
-		return inVoid || !aboveTxIsAir;
+		return !aboveTxIsAir;
 	}
 }